refactor: centralize sql helpers
